All business logic has been implemented. Methods have been added to some repositories. DatabaseConfig has been improved to work with all types of keys. Some bugs have also been fixed.

+++ ShoeWorkshop/src/main/java/org/example/entities/ReceiptItem.java
@@ -13,10 +13,10 @@ public class ReceiptItem {
     private int itemId; // Обязательно
     private int receiptId; // Обязательно
     private int favourId; // Обязательно
-    private int price; // Обязательно
+    private double price; // Обязательно
     private int quantity; // Обязательно
     
-    public ReceiptItem(int itemId, int receiptId, int favourId, int price, int quantity) {
+    public ReceiptItem(int itemId, int receiptId, int favourId, double price, int quantity) {
         this.itemId = itemId;
         this.receiptId = receiptId;
         this.favourId = favourId;
